Simplifaction refactor
