+++ action.yml
@@ -3,7 +3,7 @@ description: 'Set the context of a target Kubernetes cluster and export the kube
 inputs:
    # Please ensure you have used azure/login in the workflow before this action
    cluster-type:
-      description: 'Acceptable values: generic or arc'
+      description: 'Acceptable values: generic, arc or aks'
       required: false
       default: 'generic'
    method:
@@ -31,18 +31,10 @@ inputs:
    cluster-name:
       description: 'Azure connected cluster name. Required for az cli approach'
       required: false
-   use-kubelogin:
-      description: 'Enables kubelogin for non-admin user scenario. Required for non-admin az cli approach. Acceptable values: true or false'
-      default: 'false'
-      required: false
    admin:
-      description: 'Get cluster admin credentials. Required for az cli approach. Acceptable values: true or false'
+      description: 'Get cluster admin credentials. Required for aks clusters. Acceptable values: true or false'
       default: 'false'
       required: false
-   use-az-set-context:
-      description: 'Use az cli to get kubeconfig and set context. Acceptable values: true or false'
-      default: 'false'
-      required: true
 branding:
    color: 'blue'
 runs:
